Move left-related parser into the `left` module
  * Also fixed array index expression not handling selects
