+++ rustre-parser/src/parser.rs
@@ -236,9 +236,12 @@ fn parse_equation_equals<'slice, 'src>(input: Input<'slice, 'src>) -> IResult<'s
         EqualsEquationNode,
         join((
             alt((
-                join((parse_left, expect(t(Equal), "missing `=` in equation"))),
                 join((
-                    expect(parse_left, "missing left operand in equation"),
+                    left::parse_left,
+                    expect(t(Equal), "missing `=` in equation"),
+                )),
+                join((
+                    expect(left::parse_left, "missing left operand in equation"),
                     t(Equal),
                 )),
             )),
@@ -252,64 +255,7 @@ fn parse_equation_equals<'slice, 'src>(input: Input<'slice, 'src>) -> IResult<'s
 
 // Ebnf group LeftRules
 
-pub fn parse_left<'slice, 'src>(input: Input<'slice, 'src>) -> IResult<'slice, 'src> {
-    node(
-        LeftNode,
-        alt((
-            many_delimited(success, parse_left_item, t(Comma), peek(t(Equal))),
-            many_delimited(t(OpenPar), parse_left_item, t(Comma), t(ClosePar)),
-        )),
-    )(input)
-}
-
-pub fn parse_left_item<'slice, 'src>(input: Input<'slice, 'src>) -> IResult<'slice, 'src> {
-    fold_many1(
-        parse_id_any,
-        alt((
-            map(join((t(Dot), parse_id_any)), |c| (c, LeftFieldAccessNode)),
-            map(
-                join((
-                    t(OpenBracket),
-                    expect(
-                        alt((expression::parse_expression, parse_select)),
-                        "expected expression or select",
-                    ),
-                    t(CloseBracket),
-                )),
-                |c| (c, LeftTableAccessNode),
-            ),
-        )),
-        |a, (b, n)| (a + b).into_node(n),
-    )(input)
-}
-
-pub fn parse_select<'slice, 'src>(input: Input<'slice, 'src>) -> IResult<'slice, 'src> {
-    node(
-        SelectNode,
-        alt((
-            expression::parse_expression_15,
-            t(CDots),
-            expect(
-                expression::parse_expression_15,
-                "expected second operand in step expression",
-            ),
-            parse_step,
-        )),
-    )(input)
-}
-
-pub fn parse_step<'slice, 'src>(input: Input<'slice, 'src>) -> IResult<'slice, 'src> {
-    opt(node(
-        StepNode,
-        join((
-            t(Step),
-            expect(
-                expression::parse_expression_16,
-                "expected expression after `step`",
-            ),
-        )),
-    ))(input)
-}
+pub mod left;
 
 // Ebnf group ExpressionRules
